# docs/swagger.yaml
definitions:
  database.Datasource:
    properties:
      available:
        type: string
      did:
        type: string
      host:
        type: string
      name:
        type: string
      type:
        type: string
    type: object
  database.UserAuth:
    properties:
      address:
        type: string
      api_key:
        type: string
    type: object
  datasources.DatasourceReq:
    properties:
      host:
        type: string
      name:
        type: string
      type:
        type: string
    type: object
info:
  contact:
    name: Databroker Github Repo
    url: https://github.com/databrokerglobal/dxc
  description: Data eXchange Controller API
  license:
    name: License details
    url: https://github.com/databrokerglobal/dxc/blob/master/dbdao-license.txt
  title: DXC
  version: "1.0"
paths:
  /datasource:
    post:
      consumes:
      - application/json
      description: Create datasource
      parameters:
      - description: Datasource
        in: body
        name: datasource
        required: true
        schema:
          $ref: '#/definitions/datasources.DatasourceReq'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            type: string
        "400":
          description: Error creating datasource
          schema:
            type: string
      summary: Create datasource
      tags:
      - datasources
  /datasource/{did}:
    get:
      description: Get one datasource given a did
      parameters:
      - description: Digital identifier of the datasource
        in: path
        name: did
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Datasource'
        "500":
          description: Error retrieving datasource from database
          schema:
            type: string
      summary: Get one datasource
      tags:
      - datasources
  /datasources:
    get:
      consumes:
      - application/json
      description: Get all datasources
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Datasource'
            type: array
        "500":
          description: Error retrieving datasources from database
          schema:
            type: string
      summary: Get all datasources
      tags:
      - datasources
  /getdata/{did}:
    get:
      consumes:
      - application/json
      description: Get the data (for users)
      parameters:
      - description: Digital identifier of the data source bought
        in: path
        name: did
        required: true
        type: string
      - description: Signed verification data
        in: query
        name: verificationdata
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "401":
          description: Request not authorized. Signature and verification data invalid
          schema:
            type: string
        "404":
          description: Datasource not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get the data (for users)
      tags:
      - data
  /user/authinfo:
    get:
      description: Get address and api key for authentication with DXS
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.UserAuth'
        "404":
          description: Not data found
          schema:
            type: string
        "500":
          description: Error getting auth info
          schema:
            type: string
      summary: Get auth info
      tags:
      - user
    post:
      description: Save address and api key for authentication with DXS
      parameters:
      - description: Address
        in: query
        name: address
        required: true
        type: string
      - description: API Key
        in: query
        name: apiKey
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Error saving auth info
          schema:
            type: string
      summary: Save auth info
      tags:
      - user
swagger: "2.0"